cis_tools/pyproject.toml
[tool.poetry]
name = "cis_tools"
version = "0.1.0"
description = ""
authors = ["Andrew Erickson <aerickson@mozilla.com>"]

[tool.poetry.dependencies]
python = "^3.9"
peewee = "^3.14.4"
pendulum = "^2.1.2"
requests = "^2.26.0"
peeweext = "^1.2.3"
"github3.py" = "^2.0.0"
toml = "^0.10.2"
click = "^8.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
